pyproject.toml
[tool.poetry]
name = "registry"
version = "0.1.0"
description = ""
authors = ["nazariyv <nazariy@inbox.ru>"]

[tool.poetry.dependencies]
python = "3.9.6"
eth-brownie = "^1.16.1"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
pylint = "^2.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
